chore: Rename provider to be ThemeProvider
