+++ src/test-utils/render.tsx
@@ -1,8 +1,8 @@
-import { Provider } from '../components/ui/Provider';
+import { Provider as ThemeProvider } from '../components/ui/ThemeProvider';
 import { render as rtlRender } from '@testing-library/react';
 
 export function render(ui: React.ReactNode) {
     return rtlRender(<>{ui}</>, {
-        wrapper: (props: React.PropsWithChildren) => <Provider>{props.children}</Provider>,
+        wrapper: (props: React.PropsWithChildren) => <ThemeProvider>{props.children}</ThemeProvider>,
     });
 }
